requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
fastapi==0.115.7
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
h11==0.14.0
idna==3.10
pip==24.3.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.2
starlette-graphene3==0.6.0
strawberry-graphql==0.258.0
tagosipda-be==0.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wheel==0.45.1
